Rearrange logic between local node and sync
